+++ app/storage-service/Dockerfile
@@ -13,7 +13,6 @@ COPY ./core ./core
 ENV DATABASE_URL="file:/app/data.db"
 
 RUN bun install --filter "./app/storage-service" --filter "./core/*" --frozen-lockfile --production
-RUN bun --filter "@idol-bbq-utils/storage-service" db:generate
 RUN bun --filter "@idol-bbq-utils/storage-service" build
 
 FROM base AS runner
@@ -24,7 +23,7 @@ RUN apt-get update \
 
 WORKDIR /app
 COPY --from=builder /app/app/storage-service/build/index.js /app/bin.js
-COPY --from=builder /app/app/storage-service/prisma/client /app/prisma/client
+COPY --from=builder /app/core/db/prisma/client /app/prisma/client
 
 RUN chown -R bun:bun /app/
