+++ app/tweet-forwarder/package.json
@@ -20,8 +20,6 @@
         "url": "git+https://github.com/idol-bbq-utils/idol-bbq-utils.git"
     },
     "scripts": {
-        "db:generate-client": "bunx prisma generate",
-        "db:init": "bunx prisma migrate dev --name init",
         "build": "bun build src/main.ts --outdir=./build --target=bun --minify",
         "dev": "bun src/main.ts",
         "start": "bun src/main.ts"
@@ -31,6 +29,7 @@
     },
     "dependencies": {
         "@google/generative-ai": "^0.14.1",
+        "@idol-bbq-utils/db": "workspace:^",
         "@idol-bbq-utils/forwarder": "workspace:^",
         "@idol-bbq-utils/log": "workspace:^",
         "@idol-bbq-utils/queue": "workspace:^",
