--- app/tweet-forwarder/src/db/client.ts
@@ -1,5 +1,0 @@
-import { PrismaClient, Prisma } from '../../prisma/client/index.js'
-
-const prisma = new PrismaClient()
-
-export { prisma, Prisma }
