+++ app/tweet-forwarder/src/managers/forwarder-manager.ts
@@ -5,8 +5,8 @@ import EventEmitter from 'events'
 import { BaseCompatibleModel, sanitizeWebsites, TaskScheduler } from '@/utils/base'
 import type { AppConfig, QueueModeConfig } from '@/types'
 import { Platform, type MediaType, type TaskType } from '@idol-bbq-utils/spider/types'
-import DB from '@/db'
-import type { Article, ArticleWithId, DBFollows } from '@/db'
+import DB from '@idol-bbq-utils/db'
+import type { Article, ArticleWithId, DBFollows } from '@idol-bbq-utils/db'
 import { BaseForwarder, getForwarder } from '@idol-bbq-utils/forwarder'
 import { type Media, type MediaTool, MediaToolEnum } from '@/types/media'
 import type {
