+++ app/tweet-forwarder/src/managers/spider-manager.ts
@@ -13,10 +13,10 @@ import { BaseSpider } from '@idol-bbq-utils/spider'
 import { TranslatorProvider } from '@idol-bbq-utils/translator'
 import { translatorRegistry } from '@idol-bbq-utils/translator'
 import { pRetry } from '@idol-bbq-utils/utils'
-import DB from '@/db'
-import type { Article } from '@/db'
+import DB from '@idol-bbq-utils/db'
+import type { Article } from '@idol-bbq-utils/db'
 import { RETRY_LIMIT } from '@/config'
-import { delay } from '@/utils/time'
+import { delay } from '@idol-bbq-utils/utils/time'
 import { shuffle } from 'lodash'
 import crypto from 'crypto'
 import { QueueManager, QueueName, generateJobId, getLockKey, acquireLock, releaseLock } from '@idol-bbq-utils/queue'
