+++ core/db/.gitignore
@@ -1,0 +1,2 @@
+# Prisma Client
+prisma/client
