+++ core/db/src/client.ts
@@ -1,0 +1,5 @@
+import { PrismaClient, Prisma } from '../prisma/client'
+
+const prisma = new PrismaClient()
+
+export { prisma, Prisma }
