+++ core/utils/src/index.ts
@@ -1,3 +1,4 @@
 export * from './p-retry'
 export * from './is-network-error'
 export * from './media'
+export * from './time'
