+++ core/utils/src/time.ts
@@ -1,18 +1,18 @@
 import dayjs, { type ManipulateType } from 'dayjs'
 
-async function delay(time: number) {
+export async function delay(time: number): Promise<void> {
     return new Promise((resolve) => {
         setTimeout(() => {
-            resolve(true)
+            resolve()
         }, time)
     })
 }
 
-function formatTime(time: number) {
+export function formatTime(time: number): string {
     return dayjs.unix(time).format('YYYY-MM-DD HH:mmZ')
 }
 
-function getSubtractTime(time: number, offset: string) {
+export function getSubtractTime(time: number, offset: string): number {
     const match = offset.match(/(\d+)([a-zA-Z]+)/)
     if (match === null || !match[1] || !match[2]) throw new Error('Invalid offset format')
     return dayjs
@@ -21,4 +21,6 @@ function getSubtractTime(time: number, offset: string) {
         .unix()
 }
 
-export { delay, formatTime, getSubtractTime }
+export function isStringArrayArray(arr: string[] | string[][]): arr is Array<[string, string]> {
+    return arr.length > 0 && Array.isArray(arr[0])
+}
